+++ apps/daemon/src/runtimes/capabilities.ts
@@ -1,0 +1,2 @@
+// @ts-nocheck
+export const agentCapabilities = new Map();
