+++ apps/daemon/src/runtimes/defs/kimi.ts
@@ -1,0 +1,26 @@
+// @ts-nocheck
+import { detectAcpModels, DEFAULT_MODEL_OPTION } from './shared.js';
+
+export const kimiAgentDef = {
+    id: 'kimi',
+    name: 'Kimi CLI',
+    bin: 'kimi',
+    versionArgs: ['--version'],
+    fetchModels: async (resolvedBin, env) =>
+      detectAcpModels({
+        bin: resolvedBin,
+        args: ['acp'],
+        env,
+        timeoutMs: 15_000,
+        defaultModelOption: DEFAULT_MODEL_OPTION,
+      }),
+    fallbackModels: [
+      DEFAULT_MODEL_OPTION,
+      { id: 'kimi-k2-turbo-preview', label: 'kimi-k2-turbo-preview' },
+      { id: 'moonshot-v1-8k', label: 'moonshot-v1-8k' },
+      { id: 'moonshot-v1-32k', label: 'moonshot-v1-32k' },
+    ],
+    buildArgs: () => ['acp'],
+    streamFormat: 'acp-json-rpc',
+    mcpDiscovery: 'mature-acp',
+};
