+++ apps/daemon/src/runtimes/mcp.ts
@@ -1,0 +1,12 @@
+// @ts-nocheck
+export function buildLiveArtifactsMcpServersForAgent(def, { enabled = true, command = 'od', argsPrefix = [] } = {}) {
+  if (!enabled || def?.mcpDiscovery !== 'mature-acp') return [];
+  return [
+    {
+      name: 'open-design-live-artifacts',
+      command,
+      args: [...argsPrefix, 'mcp', 'live-artifacts'],
+      env: [],
+    },
+  ];
+}
